requirements.txt
pysoundfile